Refactor rename QuizViewControllerBuilder
